updated chair and dean fetching
updated TOS signature display
updated TOS edit

+++ app/Http/Controllers/Chairperson/ChairPOEController.php
@@ -18,6 +18,8 @@ class ChairPOEController extends Controller
         $chairperson = UserRole::where('user_id', Auth::id())
             ->where('entity_type', '=', 'Department')
             ->where('role_id', '=', Roles::where('role_name', 'Chairperson')->value('role_id'))
+            ->whereNotNull('entity_id')
+            ->orderByDesc('updated_at')
             ->firstOrFail();
         
         if ($chairperson) {
@@ -28,10 +30,10 @@ class ChairPOEController extends Controller
             $poes = POE::join('departments', 'poes.department_id', '=', 'departments.department_id')
                 ->join('user_roles', 'departments.department_id', '=', 'user_roles.entity_id')
                 ->where('user_roles.entity_type', '=', 'Department')
+                ->where('user_roles.entity_id', $department_id)
                 ->where('user_roles.role_id', '=', Roles::where('role_name', 'Chairperson')->value('role_id'))
                 // ->where('user_roles.start_validity', '<=', $today)
                 // ->where('user_roles.end_validity', '>=', $today)
-                ->where('user_roles.user_id', Auth::user()->id)
                 ->orderBy('poes.poe_code', 'asc')
                 ->select('departments.*', 'poes.*')
                 ->get();
@@ -49,6 +51,7 @@ class ChairPOEController extends Controller
     {
         $user = Auth::user();
         $notifications = $user->notifications;
+        
         return view('Chairperson.POE.poeCreate', compact('notifications'));
     }
     public function storePoe(Request $request)
@@ -56,40 +59,57 @@ class ChairPOEController extends Controller
         $chairperson = UserRole::where('user_id', Auth::id())
             ->where('entity_type', '=', 'Department')
             ->where('role_id', '=', Roles::where('role_name', 'Chairperson')->value('role_id'))
+            ->whereNotNull('entity_id')
+            ->orderByDesc('updated_at')
             ->firstOrFail();
-
-        $department_id = $chairperson->entity_id;
 
         $validatedData = $request->validate([
             'poe_code.*' => 'required',
             'poe_description.*' => 'required',
         ]);
-        foreach ($validatedData['poe_code'] as $key => $poe_code) {
-            $poe = new POE();
-            $poe->department_id = $department_id;
-            $poe->poe_code = $poe_code;
-            $poe->poe_description = $validatedData['poe_description'][$key];
-            $poe->save();
+
+        if ($chairperson) {
+            $department_id = $chairperson->entity_id;
+
+            foreach ($validatedData['poe_code'] as $key => $poe_code) {
+                $poe = new POE();
+                $poe->department_id = $department_id;
+                $poe->poe_code = $poe_code;
+                $poe->poe_description = $validatedData['poe_description'][$key];
+                $poe->save();
+            }
+            
+            return redirect()->route('chairperson.poe')->with('success', 'POE created successfully.');
+
+        } else {
+            return redirect()->back()->with('error', 'Something went wrong when retrieving Chairperson Details');
         }
 
-        return redirect()->route('chairperson.poe')->with('success', 'POE created successfully.');
     }
     public function editPoe($poe_id)
     {
-        $poes = POE::join('departments', 'poes.department_id', '=', 'departments.department_id')
-            ->join('user_roles', 'departments.department_id', '=', 'user_roles.entity_id')
-            ->where('user_roles.user_id', '=', Auth::id())
-            ->where('user_roles.entity_type', '=', 'Department')
-            ->where('user_roles.role_id', '=', Roles::where('role_name', 'Chairperson')->value('role_id'))
-            ->select('departments.*', 'poes.*')
-            ->get();
-            
         $chairperson = UserRole::where('user_id', Auth::id())
             ->where('entity_type', '=', 'Department')
             ->where('role_id', '=', Roles::where('role_name', 'Chairperson')->value('role_id'))
+            ->whereNotNull('entity_id')
+            ->orderByDesc('updated_at')
             ->firstOrFail();
 
-        $department_id = $chairperson->entity_id;
+        if ($chairperson) { 
+            $department_id = $chairperson->entity_id;
+            $poes = POE::join('departments', 'poes.department_id', '=', 'departments.department_id')
+                ->join('user_roles', 'departments.department_id', '=', 'user_roles.entity_id')
+                ->where('user_roles.user_id', '=', Auth::id())
+                ->where('user_roles.entity_type', '=', 'Department')
+                ->where('user_roles.entity_id', '=', $department_id)
+                ->where('user_roles.role_id', '=', Roles::where('role_name', 'Chairperson')->value('role_id'))
+                ->select('departments.*', 'poes.*')
+                ->get();
+
+        } else {
+            $poes = [];
+            $department_id = "";
+        }
 
         $user = Auth::user();
         $notifications = $user->notifications;
@@ -101,26 +121,33 @@ class ChairPOEController extends Controller
         $chairperson = UserRole::where('user_id', Auth::id())
             ->where('entity_type', '=', 'Department')
             ->where('role_id', '=', Roles::where('role_name', 'Chairperson')->value('role_id'))
+            ->whereNotNull('entity_id')
+            ->orderByDesc('updated_at')
             ->firstOrFail();
-
-        $department_id = $chairperson->entity_id;
 
         $validatedData = $request->validate([
             'poe_code.*' => 'required',
             'poe_description.*' => 'required',
         ]);
 
-        POE::where('department_id', $department_id)->delete();
+        if ($chairperson) {
+            $department_id = $chairperson->entity_id;
+            
+            POE::where('department_id', $department_id)->delete();
 
-        foreach ($validatedData['poe_code'] as $key => $poe_code) {
-            $poe = new POE();
-            $poe->department_id = $department_id;
-            $poe->poe_code = $poe_code;
-            $poe->poe_description = $validatedData['poe_description'][$key];
-            $poe->save();
+            foreach ($validatedData['poe_code'] as $key => $poe_code) {
+                $poe = new POE();
+                $poe->department_id = $department_id;
+                $poe->poe_code = $poe_code;
+                $poe->poe_description = $validatedData['poe_description'][$key];
+                $poe->save();
+            }
+            
+            return redirect()->route('chairperson.poe')->with('success', 'POE updated successfully.');
+
+        } else {
+            return redirect()->back()->with('error', 'Something went wrong when retrieving Chairperson Details');
         }
-
-        return redirect()->route('chairperson.poe')->with('success', 'POE updated successfully.');
     }
     public function destroyPoe($poe_id)
     {
